[type]:   fix [module]: data/widgets [id]:     #t-76030 [desc]:   Clean up the destruction code for stores that are bound to multiple data sources.
git-svn-id: http://code.extjs.com/svn/ext/trunk@4938 c0a82769-a6a8-4f8a-96ba-7530dc2e272d

+++ src/data/Store.js
@@ -527,14 +527,17 @@ sortInfo: {
      * Destroys the store.
      */
     destroy : function(){
-        if(this.storeId){
-            Ext.StoreMgr.unregister(this);
+        if(!this.isDestroyed){
+            if(this.storeId){
+                Ext.StoreMgr.unregister(this);
+            }
+            this.clearData();
+            this.data = null;
+            Ext.destroy(this.proxy);
+            this.reader = this.writer = null;
+            this.purgeListeners();
+            this.isDestroyed = true;
         }
-        this.clearData();
-        this.data = null;
-        Ext.destroy(this.proxy);
-        this.reader = this.writer = null;
-        this.purgeListeners();
     },
 
     /**
